+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+Version `dev`_
+===============
+**Date**: unreleased
+
+* CLI:
+
+  * use structlog logging library
+
 Version `0.3.0`_
 ================
 **Date**: September 06, 2019
@@ -51,3 +59,4 @@ Version `0.2.0`_
 .. _`0.2.1`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.2.0...v0.2.1
 .. _`0.2.2`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.2.1...v0.2.2
 .. _`0.3.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.2.2...v0.3.0
+.. _`dev`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.3.0...HEAD